+++ VRRS_Project/Components/Main_Components/Login_Components/Join3.js
@@ -1,9 +1,9 @@
-import React, { useState, useEffect } from 'react';
-import { View, Text, StyleSheet, ScrollView, TouchableOpacity, Alert } from 'react-native';
+import React, { useState } from 'react';
+import { View, Text, StyleSheet, TextInput } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
+import { Picker } from '@react-native-picker/picker';
 
 import Octicons from '@expo/vector-icons/Octicons';
-import MaterialIcons from '@expo/vector-icons/MaterialIcons';
 import MaterialCommunityIcons from '@expo/vector-icons/MaterialCommunityIcons';
 
 import { Gray_theme, Main_theme } from "../../../assets/styles/Theme_Colors";
@@ -14,57 +14,17 @@ import useTabBarVisibility from "../../../assets/styles/ReuseComponents/useTabBa
 export default function Join3({ navigation }) {
     useTabBarVisibility(false);
 
-    const [allAgree, setAllAgree] = useState(false);
-    const [termsAgree, setTermsAgree] = useState(false);
-    const [privacyAgree, setPrivacyAgree] = useState(false);
+    const [nickname, setNickname] = useState('');
+    const [vegType, setVegType] = useState('none');
+    const [isNicknameTouched, setIsNicknameTouched] = useState(false);
 
     const handleConfirm = () => {
-        // 모든 체크박스가 체크되었는지 확인
-        if (termsAgree && privacyAgree) {
+        if (!nickname.trim() || vegType === 'none') {
+            alert('입력한 정보를 확인해주세요.');
+        } else {
             navigation.navigate('Joinr');
-        } else {
-            Alert.alert('필수 항목에 모두 동의해 주세요.'); // 경고 메시지
         }
     };
-
-    const handleAllAgree = () => {
-        const newState = !allAgree;
-        setAllAgree(newState);
-        setTermsAgree(newState);
-        setPrivacyAgree(newState);
-    };
-
-    const toggleTermsAgree = () => {
-        setTermsAgree((prev) => {
-            const newState = !prev;
-            if (newState && privacyAgree) {
-                setAllAgree(true);
-            } else {
-                setAllAgree(false);
-            }
-            return newState;
-        });
-    };
-
-    const togglePrivacyAgree = () => {
-        setPrivacyAgree((prev) => {
-            const newState = !prev;
-            if (newState && termsAgree) {
-                setAllAgree(true);
-            } else {
-                setAllAgree(false);
-            }
-            return newState;
-        });
-    };
-
-    useEffect(() => {
-        if (termsAgree && privacyAgree) {
-            setAllAgree(true);
-        } else {
-            setAllAgree(false);
-        }
-    }, [termsAgree, privacyAgree]);
 
     return (
         <SafeAreaView style={styles.container}>
@@ -77,7 +37,7 @@ export default function Join3({ navigation }) {
             {/* 단계 섹션 */}
             <View style={styles.stepContainer}>
                 <View style={{ alignItems: 'center', justifyContent: 'center', }}>
-                    <Text style={styles.stepHeader}> 계정생성</Text>
+                    <Text style={styles.stepHeader}>정보입력</Text>
                 </View>
                 <View style={styles.stepIndicator1}>
                     <Octicons name="dot-fill" size={24} color="#E0E0E0" />
@@ -87,71 +47,54 @@ export default function Join3({ navigation }) {
                     <MaterialCommunityIcons name="numeric-3-circle" size={24} color="#468585" />
                 </View>
                 <View style={styles.stepIndicator2}>
-                    <Text style={styles.stepText2}>계정생성</Text>
+                    <Text style={styles.stepText2}>약관동의</Text>
                     <Octicons name="kebab-horizontal" size={24} color="white" />
-                    <Text style={styles.stepText2}>정보입력</Text>
+                    <Text style={styles.stepText2}>계정생성 </Text>
                     <Octicons name="kebab-horizontal" size={24} color="white" />
-                    <Text style={styles.stepText1}>약관동의  </Text>
+                    <Text style={styles.stepText1}>정보입력  </Text>
                 </View>
             </View>
-
-            {/* 전체 동의 체크박스 */}
-            <View style={styles.checkBoxAContainer}>
-                <TouchableOpacity onPress={handleAllAgree}>
-                    <View>
-                        <MaterialIcons name={allAgree ? "check-box" : "check-box-outline-blank"} size={24} color="#BDBDBD" />
-                    </View>
-                </TouchableOpacity>
-                <Text style={{ fontFamily: 'Pretendard-Medium', fontSize: 16, marginLeft: 8 }}>회원가입 약관에 모두 동의합니다.</Text>
+            {/* 닉네임 입력란 */}
+            <View style={styles.inputContainer}>
+                <Text style={styles.label}>닉네임 *</Text>
+                <TextInput
+                    style={styles.input}
+                    placeholder="닉네임을 입력하세요"
+                    value={nickname}
+                    onChangeText={(text) => {
+                        setNickname(text);
+                        setIsNicknameTouched(true); // 닉네임 입력 필드가 터치됨
+                    }}
+                />
+                {isNicknameTouched && nickname.trim() === '' && (
+                    <Text style={styles.warningText}>반드시 입력해야 하는 정보입니다.</Text>
+                )}
+                {nickname.trim() && (
+                    <Text style={styles.helperText}>사용할 수 있는 닉네임입니다.</Text>
+                )}
             </View>
 
-            {/* 이용약관 동의 */}
-            <View style={styles.checkBoxContainer}>
-                <View style={styles.checkBoxLabel}>
-                    <TouchableOpacity onPress={toggleTermsAgree}>
-                        <View>
-                            <MaterialIcons name={termsAgree ? "check-box" : "check-box-outline-blank"} size={24} color="#BDBDBD" />
-                        </View>
-                    </TouchableOpacity>
-                    <Text style={styles.checkBoxText}>이용약관 동의 
-                        <Text style={{ color: 'red', fontSize: 12 }}>  (필수)</Text>
-                    </Text>
+            {/* 채식 유형 선택 */}
+            <View style={styles.inputContainer}>
+                <Text style={styles.label}>채식 유형 *</Text>
+                <View style={styles.pickerContainer}>
+                    <Picker
+                        selectedValue={vegType}
+                        onValueChange={(itemValue) => setVegType(itemValue)}
+                        style={styles.picker}
+                    >
+                        <Picker.Item label="채식 유형을 선택해주세요." value="none" />
+                        <Picker.Item label="비건" value="vegan" />
+                        <Picker.Item label="락토 베지테리언" value="lacto" />
+                        <Picker.Item label="오보 베지테리언" value="ovo" />
+                        <Picker.Item label="락토 오보 베지테리언" value="lacto_ovo" />
+                        <Picker.Item label="페스코 베지테리언" value="pescatarian" />
+                        <Picker.Item label="폴로 베지테리언" value="pollotarian" />
+                    </Picker>
                 </View>
-                <View style={styles.agreementBox}>
-                    <ScrollView style={styles.agreementScroll}>
-                        <Text style={styles.agreementText}>
-                            제1조(목적) 이 약관은 회사(전자상거래 사업자가 운영하는 인터넷 사이버
-                            몰이용자 간의 권리와 의무에 관한 내용을 규정합니다. ...
-                        </Text>
-                    </ScrollView>
-                </View>
-            </View>
-
-            {/* 개인정보 수집 동의 */}
-            <View style={styles.checkBoxContainer}>
-                <View style={styles.checkBoxLabel}>
-                    <TouchableOpacity onPress={togglePrivacyAgree}>
-                        <View>
-                            <MaterialIcons name={privacyAgree ? "check-box" : "check-box-outline-blank"} size={24} color="#BDBDBD" />
-                        </View>
-                    </TouchableOpacity>
-                    <Text style={styles.checkBoxText}>개인 정보 수집 및 이용 동의
-                        <Text style={{ color: 'red', fontSize: 12 }}>  (필수)</Text>
-                    </Text>
-                </View>
-                <View style={styles.agreementBox}>
-                    <ScrollView style={styles.agreementScroll}>
-                        <Text style={styles.agreementText}>
-                            [제1] 개인정보 처리방침{"\n"}{"\n"}
-                            1. 목적{"\n"}
-                            2. 개인정보의 수집에 대한 동의{"\n"}
-                            3. 개인정보의 수집 및 이용 목적{"\n"}
-                            4. 흐아아{'\n'}
-                            5. 흑흑{'\n'}
-                            ...
-                        </Text>
-                    </ScrollView>
-                </View>
+                {vegType === 'none' && (
+                    <Text style={styles.warningText}>반드시 입력해야 하는 정보입니다.</Text>
+                )}
             </View>
 
             {/* 다음 버튼 */}
@@ -170,19 +113,6 @@ const styles = StyleSheet.create({
     stepContainer: {
         paddingTop: 24,
         paddingBottom: 24,
-    },
-    checkBoxAContainer: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        paddingHorizontal: 16,
-        paddingBottom: 16,
-        marginBottom: 16,
-        borderBottomWidth: 1,
-        borderBottomColor: Gray_theme.gray_30,
-    },
-    checkBoxContainer: {
-        paddingHorizontal: 16,
-        marginBottom: 16,
     },
     stepHeader: {
         fontSize: 24,
@@ -210,29 +140,46 @@ const styles = StyleSheet.create({
         fontSize: 14,
         color: Gray_theme.gray_30,
     },
-    checkBoxLabel: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        marginBottom: 8,
+    inputContainer: {
+        position: 'relative',
+        marginBottom: 16,
+        paddingHorizontal: 16,
     },
-    checkBoxText: {
-        fontFamily: 'Pretendard-Regular',
+    label: {
         fontSize: 16,
+        fontFamily: 'Pretendard-Medium',
         marginLeft: 8,
     },
-    agreementBox: {
-        borderWidth: 1,
-        borderColor: Gray_theme.gray_30,
-        borderRadius: 3,
-        height: 120,
-    },
-    agreementScroll: {
-        padding: 16,
-    },
-    agreementText: {
-        fontFamily: 'Pretendard-Regular',
+    input: {
+        height: 42,
+        paddingHorizontal: 8,
+        paddingVertical: 4,
+        borderBottomWidth: 1,
+        borderBottomColor: Gray_theme.gray_40,
+        fontFamily: "Pretendard-Regular",
         fontSize: 14,
-        color: Gray_theme.balck,
+    },
+    helperText: {
+        fontSize: 12,
+        fontFamily: 'Pretendard-Regular',
+        color: Main_theme.main_50,
+        marginTop: 4,
+    },
+    warningText: {
+        fontSize: 12,
+        fontFamily: 'Pretendard-Regular',
+        color: Main_theme.main_reverse,
+        marginTop: 4,
+    },
+    pickerContainer: {
+        borderWidth: 1,
+        borderColor: Gray_theme.gray_40,
+        borderRadius: 8,
+        marginTop: 8,
+    },
+    picker: {
+        height: 48,
+        width: '100%',
     },
     button: {
         position: "absolute",
