+++ package.json
@@ -9,7 +9,7 @@
     "@react-navigation/material-top-tabs": "^6.6.14",
     "@react-navigation/native": "^6.1.18",
     "@react-navigation/native-stack": "^6.11.0",
-    "expo": "~51.0.37",
+    "expo": "~51.0.38",
     "expo-camera": "^15.0.16",
     "expo-image-manipulator": "^12.0.5",
     "expo-image-picker": "^15.0.7",
